backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
PyPDF2==3.0.1
google-generativeai==0.3.2
python-dotenv==1.0.0
transformers==4.35.2
spacy==3.7.2
